feat: Move Loki components to integrations/loki/
