+++ src/fapilog/sinks/loki.py
@@ -11,14 +11,14 @@ except ImportError:
     httpx = None
 
 from .._internal.error_handling import StandardSinkErrorHandling
-from .._internal.loki_http_client import LokiHttpClient
-from .._internal.loki_payload_formatter import LokiPayloadFormatter
 from ..async_components.queue.batch_manager import BatchManager
 from ..exceptions import (
     ConfigurationError,
     SinkConfigurationError,
     SinkErrorContextBuilder,
 )
+from ..integrations.loki.client import LokiHttpClient
+from ..integrations.loki.formatter import LokiPayloadFormatter
 from .base import Sink
 
 if TYPE_CHECKING:
